Adds missing methods into the CMS API and adds default implementations in the ActiveMQ-CPP code for them that throw UnsupportedOperationExceptions.  Cleans up a bunch of code in preperation for the release, fixes warnings and removes debug code.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@775980 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/test/decaf/net/URISyntaxExceptionTest.cpp
@@ -38,5 +38,5 @@ void URISyntaxExceptionTest::test() {
     URISyntaxException e2( __FILE__, __LINE__, "str", "problem");
     CPPUNIT_ASSERT_MESSAGE("returned incorrect reason", e2.getReason() == "problem" );
     CPPUNIT_ASSERT_MESSAGE("returned incorrect input", e2.getInput() == "str" );
-    CPPUNIT_ASSERT_MESSAGE("returned incorrect index", -1 == e2.getIndex());
+    CPPUNIT_ASSERT_MESSAGE("returned incorrect index", -1 == (int)e2.getIndex());
 }
